Refactor: Move Settings window to HunterPie module
