+++ HunterPie/GUI/Parts/Settings/Views/SettingHost.xaml
@@ -1,17 +1,19 @@
-﻿<UserControl x:Class="HunterPie.UI.Controls.Settings.SettingHost"
+﻿<UserControl x:Class="HunterPie.GUI.Parts.Settings.Views.SettingHost"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:HunterPie.UI.Controls.Settings"
-             xmlns:converters="clr-namespace:HunterPie.UI.Architecture.Converters"
-             xmlns:native="clr-namespace:HunterPie.UI.Controls.Buttons"
-             xmlns:textbox="clr-namespace:HunterPie.UI.Controls.TextBox"
+             xmlns:local="clr-namespace:HunterPie.GUI.Parts.Settings.Views"
+             xmlns:converters="clr-namespace:HunterPie.UI.Architecture.Converters;assembly=HunterPie.UI"
+             xmlns:native="clr-namespace:HunterPie.UI.Controls.Buttons;assembly=HunterPie.UI"
+             xmlns:textbox="clr-namespace:HunterPie.UI.Controls.TextBox;assembly=HunterPie.UI"
              xmlns:client="clr-namespace:HunterPie.Core.Client;assembly=HunterPie.Core"
-             xmlns:viewer="clr-namespace:HunterPie.UI.Controls.Viewer"
-             xmlns:vm="clr-namespace:HunterPie.UI.Controls.Settings.ViewModel"
-             xmlns:loading="clr-namespace:HunterPie.UI.Controls.Loading"
-             xmlns:loc="clr-namespace:HunterPie.UI.Architecture.Assets"
+             xmlns:viewer="clr-namespace:HunterPie.UI.Controls.Viewer;assembly=HunterPie.UI"
+             xmlns:vm="clr-namespace:HunterPie.GUI.Parts.Settings.ViewModels"
+             xmlns:loading="clr-namespace:HunterPie.UI.Controls.Loading;assembly=HunterPie.UI"
+             xmlns:loc="clr-namespace:HunterPie.UI.Architecture.Assets;assembly=HunterPie.UI"
+             xmlns:viewModel="clr-namespace:HunterPie.UI.Controls.Settings.ViewModel;assembly=HunterPie.UI"
+             xmlns:settings="clr-namespace:HunterPie.UI.Controls.Settings;assembly=HunterPie.UI"
              mc:Ignorable="d"
              d:DesignHeight="450" 
              d:DesignWidth="800"
@@ -40,8 +42,8 @@
         <converters:ObjectToUiConverter x:Key="ObjectToUiConverter"/>
         <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter"/>
         <converters:InverseBooleanToVisibilityConverter x:Key="InverseBooleanToVisibilityConverter"/>
-        <vm:MockSettingHostViewModel x:Key="MockSettingHostViewModel"/>
-        <DataTemplate x:Key="PanelTemplate" DataType="{x:Type vm:ISettingElement}">
+        <viewModel:MockSettingHostViewModel x:Key="MockSettingHostViewModel"/>
+        <DataTemplate x:Key="PanelTemplate" DataType="{x:Type viewModel:ISettingElement}">
             <ItemsControl ItemsSource="{Binding Elements, Mode=OneWay}">
                 <ItemsControl.Template>
                     <ControlTemplate>
@@ -55,8 +57,8 @@
                 </ItemsControl.Template>
 
                 <ItemsControl.ItemTemplate>
-                    <DataTemplate DataType="{x:Type vm:ISettingElementType}">
-                        <local:SettingElementHost Hosted="{Binding ., Converter={StaticResource ObjectToUiConverter}, IsAsync=True}"
+                    <DataTemplate DataType="{x:Type viewModel:ISettingElementType}">
+                        <settings:SettingElementHost Hosted="{Binding ., Converter={StaticResource ObjectToUiConverter}, IsAsync=True}"
                                                   Text="{Binding Name}"
                                                   Description="{Binding Description}"
                                                   Visibility="{Binding Match, Converter={StaticResource BooleanToVisibilityConverter}}"
@@ -67,7 +69,7 @@
                 </ItemsControl.ItemTemplate>
             </ItemsControl>
         </DataTemplate>
-        <DataTemplate x:Key="HeaderTemplate" DataType="{x:Type vm:ISettingElement}">
+        <DataTemplate x:Key="HeaderTemplate" DataType="{x:Type viewModel:ISettingElement}">
             <native:TabItem Padding="10,5,10,5"
                             Title="{Binding Title}"
                             Description="{Binding Description}"
