+++ HunterPie/HunterPie.csproj
@@ -90,6 +90,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Update="GUI\Parts\Settings\Views\SettingHost.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Update="Properties\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
