+++ HunterPie/HunterPie.csproj.user
@@ -146,6 +146,9 @@
     <Page Update="GUI\Parts\Patches\Views\PatchesView.xaml">
       <SubType>Designer</SubType>
     </Page>
+    <Page Update="GUI\Parts\Settings\Views\SettingHost.xaml">
+      <SubType>Designer</SubType>
+    </Page>
     <Page Update="GUI\Parts\Sidebar\SideBarContainer.xaml">
       <SubType>Designer</SubType>
     </Page>
